merge develop into main

+++ src/main/java/br/duosilva/tech/solutions/ez/frame/generator/ms/domain/service/VideoProcessingService.java
@@ -4,10 +4,11 @@ import java.io.File;
 import java.util.List;
 
 import org.apache.commons.io.FilenameUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import br.duosilva.tech.solutions.ez.frame.generator.ms.frameworks.exception.BusinessRuleException;
 import br.duosilva.tech.solutions.ez.frame.generator.ms.infrastructure.ffmpeg.FFmpegFrameExtractor;
 import br.duosilva.tech.solutions.ez.frame.generator.ms.infrastructure.zip.ZipFileGenerator;
 
@@ -16,6 +17,7 @@ public class VideoProcessingService {
 
 	private final FFmpegFrameExtractor ffmpegFrameExtractor;
 	private final ZipFileGenerator zipFileGenerator;
+	private static final Logger LOGGER = LoggerFactory.getLogger(VideoProcessingService.class);
 
 	public VideoProcessingService(FFmpegFrameExtractor ffmpegFrameExtractor, ZipFileGenerator zipFileGenerator) {
 		this.ffmpegFrameExtractor = ffmpegFrameExtractor;
@@ -33,32 +35,19 @@ public class VideoProcessingService {
 	 * @param userId        ID do usuário
 	 * @return Arquivo .zip contendo os frames extraídos
 	 */
-	public File generateFrames(MultipartFile multipartFile, String userId) {
+	public File generateFrames(MultipartFile multipartFile) {
+		List<File> frames = ffmpegFrameExtractor.extractFrames(multipartFile);
+	    String baseName = FilenameUtils.getBaseName(multipartFile.getOriginalFilename());
+	    File zipFile = compressFrames(frames, baseName);
 
-		try {
-			// 1. Extrair frames e salvar temporariamente
-			List<File> frames = ffmpegFrameExtractor.extractFrames(multipartFile);
+	    // Clean up
+	    frames.forEach(File::delete);
+	    if (!frames.isEmpty()) {
+	        File frameDirectory = frames.get(0).getParentFile();
+	        frameDirectory.delete();
+	    }
 
-			// 2.Gerar baseName do vídeo original
-			String baseName = FilenameUtils.getBaseName(multipartFile.getOriginalFilename());
-
-			// 3. Gerar o .zip dos frames
-			File zipFile = compressFrames(frames, baseName);
-
-			// 4. Limpar arquivos de frame individuais
-			for (File frame : frames) {
-				frame.delete();
-			}
-
-			// 5. Deletar diretorio temporario (caso esteja vazio)
-			File frameDirectory = frames.get(0).getParentFile();
-			frameDirectory.delete(); // so funciona se a pasta estiver vazia
-
-			return zipFile;
-
-		} catch (Exception e) {
-			throw new BusinessRuleException("Failed to process video and generate frames: " + e.getMessage());
-		}
+	    return zipFile;
 	}
 
 	/**
@@ -71,5 +60,5 @@ public class VideoProcessingService {
 	private File compressFrames(List<File> frames, String baseName) {
 		return zipFileGenerator.generateZipFromFrames(frames, baseName);
 	}
-
+	
 }
